@@ -1,27 +1,25 @@
 name: firebase_auth_rest
 description: A platform independent Dart/Flutter Wrapper for the Firebase
   Authentication API based on REST
-version: 1.0.2
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/Skycoder42/firebase_auth_rest
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  freezed_annotation: ^0.12.0
-  http: ^0.12.2
-  json_annotation: ^3.1.1
-  logging: ^0.11.4
+  freezed_annotation: ^0.13.0-nullsafety.0
+  http: ^0.13.0-nullsafety.0
+  json_annotation: ^4.0.0-nullsafety.0
 
 dev_dependencies:
-  build_runner: ^1.11.0
-  cider: ^0.0.5
-  coverage: ^0.14.1
-  dart_pre_commit: ^1.1.4
-  freezed: ^0.12.7
-  git_hooks: ^0.0.4
-  hanzo: ^1.0.2
-  json_serializable: ^3.5.1
-  lint: ^1.3.0
-  mockito: ^4.1.4
-  test: ^1.15.3
+  build_runner: ^1.11.1 # X
+  # cider: ^0.0.5 # X
+  coverage: ^0.15.1 # X
+  dart_pre_commit: ^2.2.0-nullsafety.0
+  freezed: ^0.13.0-nullsafety.2
+  json_serializable: ^4.0.0-nullsafety.0
+  lint: ^1.5.1
+  mockito: ^5.0.0-nullsafety.5
+  test: ^1.16.0-nullsafety.19
+  tuple: ^2.0.0-nullsafety.0
