@@ -10,13 +10,13 @@ abstract class ErrorDetails with _$ErrorDetails {
   /// Default constructor
   const factory ErrorDetails({
     /// The domain in which the error occured
-    String domain,
+    String? domain,
 
     /// The reason for the error
-    String reason,
+    String? reason,
 
     /// The error message / code
-    String message,
+    String? message,
   }) = _ErrorDetails;
 
   /// JSON constructor
@@ -30,13 +30,13 @@ abstract class ErrorData with _$ErrorData {
   /// Default constructor
   const factory ErrorData({
     /// The error code
-    int code,
+    @Default(-1) int code,
 
     /// The error message
-    String message,
+    String? message,
 
     /// A list of details about this error
-    List<ErrorDetails> errors,
+    @Default(<ErrorDetails>[]) List<ErrorDetails> errors,
   }) = _ErrorData;
 
   /// JSON Constructor
@@ -48,10 +48,10 @@ abstract class ErrorData with _$ErrorData {
 @freezed
 abstract class AuthException with _$AuthException implements Exception {
   /// Default constructor
-  const factory AuthException([
+  const factory AuthException(
     /// The actual error data
     ErrorData error,
-  ]) = _AuthException;
+  ) = _AuthException;
 
   /// JSON Constructor
   factory AuthException.fromJson(Map<String, dynamic> json) =>
