@@ -12,10 +12,10 @@ abstract class ProviderUserInfo with _$ProviderUserInfo {
   /// Default constructor
   const factory ProviderUserInfo({
     /// The linked provider ID (e.g. "google.com" for the Google provider).
-    String providerId,
+    required String providerId,
 
     /// The unique ID identifies the IdP account.
-    String federatedId,
+    required String federatedId,
   }) = _ProviderUserInfo;
 
   /// JSON constructor
