@@ -10,22 +10,22 @@ abstract class RefreshResponse with _$RefreshResponse {
   /// Default constructor
   const factory RefreshResponse({
     /// The number of seconds in which the ID token expires.
-    String expires_in,
+    required String expires_in,
     // The type of the refresh token, always "Bearer".
-    String token_type,
+    required String token_type,
 
     /// The Firebase Auth refresh token provided in the request or a new refresh
     /// token.
-    String refresh_token,
+    required String refresh_token,
 
     /// A Firebase Auth ID token.
-    String id_token,
+    required String id_token,
 
     /// The uid corresponding to the provided ID token.
-    String user_id,
+    required String user_id,
 
     /// Your Firebase project ID.
-    String project_id,
+    required String project_id,
   }) = _RefreshResponse;
 
   /// JSON constructor
