@@ -9,79 +9,79 @@ abstract class SignInResponse with _$SignInResponse {
   /// https://firebase.google.com/docs/reference/rest/auth#section-sign-in-anonymously
   const factory SignInResponse.anonymous({
     /// A Firebase Auth ID token for the newly created user.
-    String idToken,
+    required String idToken,
 
     /// Since the user is anonymous, this should be empty.
-    String email,
+    String? email,
 
     /// A Firebase Auth refresh token for the newly created user.
-    String refreshToken,
+    required String refreshToken,
 
     /// The number of seconds in which the ID token expires.
-    String expiresIn,
+    required String expiresIn,
 
     /// The uid of the newly created user.
-    String localId,
+    required String localId,
   }) = AnonymousSignInResponse;
 
   /// https://firebase.google.com/docs/reference/rest/auth#section-sign-in-with-oauth-credential
   const factory SignInResponse.idp({
     /// The unique ID identifies the IdP account.
-    String federatedId,
+    required String federatedId,
 
     /// The linked provider ID (e.g. "google.com" for the Google provider).
-    String providerId,
+    required String providerId,
 
     /// The uid of the authenticated user.
-    String localId,
+    required String localId,
 
     /// Whether the sign-in email is verified.
-    bool emailVerified,
+    @Default(false) bool emailVerified,
 
     /// The email of the account.
-    String email,
+    String? email,
 
     /// The OIDC id token if available.
-    String oauthIdToken,
+    String? oauthIdToken,
 
     /// The OAuth access token if available.
-    String oauthAccessToken,
+    String? oauthAccessToken,
 
     /// The OAuth 1.0 token secret if available.
-    String oauthTokenSecret,
+    String? oauthTokenSecret,
 
     /// The stringified JSON response containing all the IdP data corresponding
     /// to the provided OAuth credential.
-    String rawUserInfo,
+    String? rawUserInfo,
 
     /// The first name for the account.
-    String firstName,
+    String? firstName,
 
     /// The last name for the account.
-    String lastName,
+    String? lastName,
 
     /// The full name for the account.
-    String fullName,
+    String? fullName,
 
     /// The display name for the account.
-    String displayName,
+    String? displayName,
 
     /// The photo Url for the account.
-    Uri photoUrl,
+    Uri? photoUrl,
 
     /// A Firebase Auth ID token for the authenticated user.
-    String idToken,
+    required String idToken,
 
     /// A Firebase Auth refresh token for the authenticated user.
-    String refreshToken,
+    required String refreshToken,
 
     /// The number of seconds in which the ID token expires.
-    String expiresIn,
+    required String expiresIn,
 
     /// Whether another account with the same credential already exists. The
     /// user will need to sign in to the original account and then link the
     /// current credential to it.
-    bool needConfirmation,
+    @Default(false) bool needConfirmation,
   }) = IdpSignInResponse;
 
   /// Meta-Class for multiple API-Endpoints
@@ -90,92 +90,92 @@ abstract class SignInResponse with _$SignInResponse {
   /// - https://firebase.google.com/docs/reference/rest/auth#section-sign-in-email-password
   const factory SignInResponse.password({
     /// A Firebase Auth ID token for the authenticated user.
-    String idToken,
+    required String idToken,
 
     /// The email for the authenticated user.
-    String email,
+    String? email,
 
     /// A Firebase Auth refresh token for the authenticated user.
-    String refreshToken,
+    required String refreshToken,
 
     /// The number of seconds in which the ID token expires.
-    String expiresIn,
+    required String expiresIn,
 
     /// The uid of the authenticated user.
-    String localId,
+    required String localId,
 
     /// Whether the email is for an existing account.
-    bool registered,
+    @Default(false) bool registered,
   }) = PasswordSignInResponse; // TODO split into two
 
   /// https://firebase.google.com/docs/reference/rest/auth#section-verify-custom-token
   const factory SignInResponse.custom({
     /// A Firebase Auth ID token generated from the provided custom token.
-    String idToken,
+    required String idToken,
 
     /// A Firebase Auth refresh token generated from the provided custom token.
-    String refreshToken,
+    required String refreshToken,
 
     /// The number of seconds in which the ID token expires.
-    String expiresIn,
+    required String expiresIn,
 
-    /// @nodoc
-    String localId,
+    /// The uid of the authenticated user.
+    required String localId,
   }) = CustomTokenSignInResponse;
 
   /// https://firebase.google.com/docs/reference/rest/auth#section-link-with-oauth-credential
   const factory SignInResponse.linkIdp({
     /// The unique ID identifies the IdP account.
-    String federatedId,
+    required String federatedId,
 
     /// The linked provider ID (e.g. "google.com" for the Google provider).
-    String providerId,
+    required String providerId,
 
     /// The uid of the authenticated user.
-    String localId,
+    required String localId,
 
     /// Whether the signin email is verified.
-    bool emailVerified,
+    @Default(false) bool emailVerified,
 
     /// The email of the account.
-    String email,
+    String? email,
 
     /// The OIDC id token if available.
-    String oauthIdToken,
+    String? oauthIdToken,
 
     /// The OAuth access token if available.
-    String oauthAccessToken,
+    String? oauthAccessToken,
 
     /// The OAuth 1.0 token secret if available.
-    String oauthTokenSecret,
+    String? oauthTokenSecret,
 
     /// The stringified JSON response containing all the IdP data corresponding
     /// to the provided OAuth credential.
-    String rawUserInfo,
+    String? rawUserInfo,
 
     /// The first name for the account.
-    String firstName,
+    String? firstName,
 
     /// The last name for the account.
-    String lastName,
+    String? lastName,
 
     /// The full name for the account.
-    String fullName,
+    String? fullName,
 
     /// The display name for the account.
-    String displayName,
+    String? displayName,
 
     /// The photo Url for the account.
-    Uri photoUrl,
+    Uri? photoUrl,
 
     /// A Firebase Auth ID token for the authenticated user.
-    String idToken,
+    required String idToken,
 
     /// A Firebase Auth refresh token for the authenticated user.
-    String refreshToken,
+    required String refreshToken,
 
     /// The number of seconds in which the ID token expires.
-    String expiresIn,
+    required String expiresIn,
   }) = LinkIdpResponse;
 
   /// JSON constructor
