@@ -12,7 +12,7 @@ abstract class UserDataResponse with _$UserDataResponse {
   const factory UserDataResponse({
     /// The account associated with the given Firebase ID token. Check
     /// [UserData] for more details.
-    List<UserData> users,
+    @Default(<UserData>[]) List<UserData> users,
   }) = _UserDataResponse;
 
   /// JSON constructor
