@@ -5,15 +5,18 @@ import 'package:test/test.dart';
 void main() {
   test('update works as expected', () {
     final update = ProfileUpdate<int>.update(42);
-    expect(update.update, true);
-    expect(update.delete, false);
+    expect(update.isUpdate, true);
+    expect(update.isDelete, false);
     expect(update.data, 42);
+    expect(update.updateOr(55), 42);
   });
 
   test('delete works as expected', () {
     final update = ProfileUpdate<int>.delete();
-    expect(update.update, false);
-    expect(update.delete, true);
+    expect(update.isUpdate, false);
+    expect(update.isDelete, true);
     expect(update.data, null);
+    expect(update.updateOr(), null);
+    expect(update.updateOr(55), 55);
   });
 }
